requirements.txt
aiogram
pydantic
asyncpg
aiohttp
requests
tomli